TGP-3179: refactor nhasCorrenctGoodsController
