+++ app/models/outboundLink/OutboundLink.scala
@@ -153,8 +153,8 @@ object OutboundLink {
     val link: String            = "https://www.trade-tariff.service.gov.uk/xi/find_commodity"
     val linkTextKey: String     = "hasCorrectGoods.p2.linkText"
     val originatingPage: String = recordId match {
-      case Some(recordId) => controllers.routes.HasCorrectGoodsController.onSubmitUpdate(mode, recordId).url
-      case _              => controllers.routes.HasCorrectGoodsController.onSubmitCreate(mode).url
+      case Some(recordId) => controllers.hasCorrectGoods.routes.UpdateController.onSubmit(mode, recordId).url
+      case _              => controllers.hasCorrectGoods.routes.CreateController.onSubmit(mode).url
     }
   }
