+++ test/controllers/hasCorrectGoods/UpdateControllerSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.hasCorrectGoods
 
 import base.SpecBase
 import base.TestConstants.{testEori, testRecordId}
@@ -34,7 +34,7 @@ import play.api.inject.bind
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
-import queries.{CommodityQuery, CommodityUpdateQuery, LongerCommodityQuery}
+import queries.CommodityUpdateQuery
 import repositories.SessionRepository
 import services.AuditService
 import uk.gov.hmrc.auth.core.AffinityGroup
@@ -44,7 +44,7 @@ import views.html.HasCorrectGoodsView
 import java.time.Instant
 import scala.concurrent.Future
 
-class HasCorrectGoodsControllerSpec extends SpecBase with MockitoSugar {
+class UpdateControllerSpec extends SpecBase with MockitoSugar {
 
   private def onwardRoute = Call("GET", "/foo")
 
@@ -66,409 +66,11 @@ class HasCorrectGoodsControllerSpec extends SpecBase with MockitoSugar {
 
   "HasCorrectGoodsController" - {
 
-    "For create journey" - {
-      lazy val hasCorrectGoodsCreateRoute =
-        routes.HasCorrectGoodsController.onPageLoadCreate(NormalMode).url
-      lazy val onSubmitAction: Call       =
-        routes.HasCorrectGoodsController.onSubmitCreate(NormalMode)
-      val page: QuestionPage[Boolean]     = HasCorrectGoodsPage
-
-      "must return OK and the correct view for a GET" in {
-
-        val userAnswers =
-          emptyUserAnswers
-            .set(CommodityQuery, Commodity("654321", List("Description", "Other"), Instant.now, None))
-            .success
-            .value
-
-        val application = applicationBuilder(userAnswers = Some(userAnswers))
-          .build()
-
-        running(application) {
-          val request = FakeRequest(GET, hasCorrectGoodsCreateRoute)
-
-          val result = route(application, request).value
-
-          val view = application.injector.instanceOf[HasCorrectGoodsView]
-
-          status(result) mustEqual OK
-          contentAsString(result) mustEqual view(
-            form,
-            Commodity("654321", List("Description", "Other"), Instant.now, None),
-            onSubmitAction,
-            NormalMode,
-            None
-          )(
-            request,
-            messages(application)
-          ).toString
-        }
-      }
-
-      "must redirect on GET to JourneyRecovery Page if user doesn't have commodity answer" in {
-
-        val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
-
-        running(application) {
-          val request = FakeRequest(GET, hasCorrectGoodsCreateRoute)
-
-          val result = route(application, request).value
-
-          status(result) mustEqual SEE_OTHER
-          redirectLocation(result).value mustEqual controllers.problem.routes.JourneyRecoveryController.onPageLoad().url
-        }
-      }
-
-      "must populate the view correctly on a GET when the question has previously been answered" in {
-
-        val commodity   = Commodity("654321", List("Description"), Instant.now, None)
-        val userAnswers = emptyUserAnswers
-          .set(CommodityQuery, commodity)
-          .success
-          .value
-          .set(page, true)
-          .success
-          .value
-
-        val application = applicationBuilder(userAnswers = Some(userAnswers)).build()
-
-        running(application) {
-          val request = FakeRequest(GET, hasCorrectGoodsCreateRoute)
-
-          val view = application.injector.instanceOf[HasCorrectGoodsView]
-
-          val result = route(application, request).value
-
-          status(result) mustEqual OK
-          contentAsString(result) mustEqual view(form.fill(true), commodity, onSubmitAction, NormalMode, None)(
-            request,
-            messages(application)
-          ).toString
-        }
-      }
-
-      "must redirect to the next page when valid data is submitted" in {
-
-        val mockSessionRepository = mock[SessionRepository]
-
-        when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
-
-        val application =
-          applicationBuilder(userAnswers = Some(emptyUserAnswers))
-            .overrides(
-              bind[Navigation].toInstance(new FakeNavigation(onwardRoute)),
-              bind[SessionRepository].toInstance(mockSessionRepository)
-            )
-            .build()
-
-        running(application) {
-          val request =
-            FakeRequest(POST, hasCorrectGoodsCreateRoute)
-              .withFormUrlEncodedBody(("value", "true"))
-
-          val result = route(application, request).value
-
-          status(result) mustEqual SEE_OTHER
-          redirectLocation(result).value mustEqual onwardRoute.url
-        }
-      }
-
-      "must redirect on POST to JourneyRecovery Page if user doesn't have commodity answer" in {
-
-        val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
-
-        running(application) {
-          val request =
-            FakeRequest(POST, hasCorrectGoodsCreateRoute)
-              .withFormUrlEncodedBody(("value", ""))
-
-          val result = route(application, request).value
-
-          status(result) mustEqual SEE_OTHER
-          redirectLocation(result).value mustEqual controllers.problem.routes.JourneyRecoveryController.onPageLoad().url
-        }
-      }
-
-      "must return a Bad Request and errors when invalid data is submitted" in {
-
-        val commodity = Commodity("654321", List("Description"), Instant.now, None)
-
-        val userAnswers =
-          emptyUserAnswers.set(CommodityQuery, commodity).success.value
-
-        val application = applicationBuilder(userAnswers = Some(userAnswers)).build()
-
-        running(application) {
-          val request =
-            FakeRequest(POST, hasCorrectGoodsCreateRoute)
-              .withFormUrlEncodedBody(("value", ""))
-
-          val boundForm = form.bind(Map("value" -> ""))
-
-          val view = application.injector.instanceOf[HasCorrectGoodsView]
-
-          val result = route(application, request).value
-
-          status(result) mustEqual BAD_REQUEST
-          contentAsString(result) mustEqual view(boundForm, commodity, onSubmitAction, NormalMode, None)(
-            request,
-            messages(application)
-          ).toString
-        }
-      }
-
-      "must redirect to Journey Recovery for a GET if no existing data is found" in {
-
-        val application = applicationBuilder(userAnswers = None).build()
-
-        running(application) {
-          val request = FakeRequest(GET, hasCorrectGoodsCreateRoute)
-
-          val result = route(application, request).value
-
-          status(result) mustEqual SEE_OTHER
-          redirectLocation(result).value mustEqual controllers.problem.routes.JourneyRecoveryController.onPageLoad().url
-        }
-      }
-
-      "must redirect to Journey Recovery for a POST if no existing data is found" in {
-
-        val application = applicationBuilder(userAnswers = None).build()
-
-        running(application) {
-          val request =
-            FakeRequest(POST, hasCorrectGoodsCreateRoute)
-              .withFormUrlEncodedBody(("value", "true"))
-
-          val result = route(application, request).value
-
-          status(result) mustEqual SEE_OTHER
-          redirectLocation(result).value mustEqual controllers.problem.routes.JourneyRecoveryController.onPageLoad().url
-        }
-      }
-    }
-
-    "for the Longer Commodity Code Journey" - {
-
-      lazy val hasCorrectGoodsRoute =
-        routes.HasCorrectGoodsController
-          .onPageLoadLongerCommodityCode(NormalMode, testRecordId)
-          .url
-      lazy val onSubmitAction: Call =
-        routes.HasCorrectGoodsController.onSubmitLongerCommodityCode(NormalMode, testRecordId)
-
-      "for a GET" - {
-
-        "must return OK and the correct view" in {
-
-          val userAnswers =
-            emptyUserAnswers
-              .set(
-                LongerCommodityQuery(testRecordId),
-                Commodity("654321", List("Description", "Other"), Instant.now, None)
-              )
-              .success
-              .value
-
-          val application = applicationBuilder(userAnswers = Some(userAnswers))
-            .build()
-
-          running(application) {
-            val request = FakeRequest(GET, hasCorrectGoodsRoute)
-
-            val result = route(application, request).value
-
-            val view = application.injector.instanceOf[HasCorrectGoodsView]
-
-            status(result) mustEqual OK
-            contentAsString(result) mustEqual view(
-              form,
-              Commodity("654321", List("Description", "Other"), Instant.now, None),
-              onSubmitAction,
-              NormalMode,
-              Some(testRecordId)
-            )(
-              request,
-              messages(application)
-            ).toString
-          }
-        }
-
-        "must redirect to JourneyRecovery Page if user doesn't have commodity answer" in {
-
-          val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
-
-          running(application) {
-            val request = FakeRequest(GET, hasCorrectGoodsRoute)
-
-            val result = route(application, request).value
-
-            status(result) mustEqual SEE_OTHER
-            redirectLocation(result).value mustEqual controllers.problem.routes.JourneyRecoveryController
-              .onPageLoad()
-              .url
-          }
-        }
-
-        "must populate the view correctly when the question has previously been answered" in {
-
-          val commodity   = Commodity("654321", List("Description"), Instant.now, None)
-          val userAnswers = emptyUserAnswers
-            .set(LongerCommodityQuery(testRecordId), commodity)
-            .success
-            .value
-            .set(HasCorrectGoodsLongerCommodityCodePage(testRecordId), true)
-            .success
-            .value
-
-          val application = applicationBuilder(userAnswers = Some(userAnswers)).build()
-
-          running(application) {
-            val request = FakeRequest(GET, hasCorrectGoodsRoute)
-
-            val view = application.injector.instanceOf[HasCorrectGoodsView]
-
-            val result = route(application, request).value
-
-            status(result) mustEqual OK
-            contentAsString(result) mustEqual view(
-              form.fill(true),
-              commodity,
-              onSubmitAction,
-              NormalMode,
-              Some(testRecordId)
-            )(
-              request,
-              messages(application)
-            ).toString
-          }
-        }
-
-        "must redirect to Journey Recovery if no existing data is found" in {
-
-          val application = applicationBuilder(userAnswers = None).build()
-
-          running(application) {
-            val request = FakeRequest(GET, hasCorrectGoodsRoute)
-
-            val result = route(application, request).value
-
-            status(result) mustEqual SEE_OTHER
-            redirectLocation(result).value mustEqual controllers.problem.routes.JourneyRecoveryController
-              .onPageLoad()
-              .url
-          }
-        }
-
-      }
-
-      "for a POST" - {
-        "must redirect to the next page when valid data is submitted" in {
-
-          val userAnswers =
-            emptyUserAnswers
-              .set(
-                LongerCommodityQuery(testRecordId),
-                Commodity("654321", List("Description", "Other"), Instant.now, None)
-              )
-              .success
-              .value
-
-          val mockSessionRepository = mock[SessionRepository]
-
-          when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
-
-          val application =
-            applicationBuilder(userAnswers = Some(userAnswers))
-              .overrides(
-                bind[Navigation].toInstance(new FakeNavigation(onwardRoute)),
-                bind[SessionRepository].toInstance(mockSessionRepository)
-              )
-              .build()
-
-          running(application) {
-            val request =
-              FakeRequest(POST, hasCorrectGoodsRoute)
-                .withFormUrlEncodedBody(("value", "true"))
-
-            val result = route(application, request).value
-
-            status(result) mustEqual SEE_OTHER
-            redirectLocation(result).value mustEqual onwardRoute.url
-          }
-
-        }
-
-        "must redirect to JourneyRecovery Page if user doesn't have commodity answer" in {
-
-          val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
-
-          running(application) {
-            val request =
-              FakeRequest(POST, hasCorrectGoodsRoute)
-                .withFormUrlEncodedBody(("value", ""))
-
-            val result = route(application, request).value
-
-            status(result) mustEqual SEE_OTHER
-            redirectLocation(result).value mustEqual controllers.problem.routes.JourneyRecoveryController
-              .onPageLoad()
-              .url
-          }
-        }
-
-        "must return a Bad Request and errors when invalid data is submitted" in {
-
-          val commodity = Commodity("654321", List("Description"), Instant.now, None)
-
-          val userAnswers =
-            emptyUserAnswers.set(LongerCommodityQuery(testRecordId), commodity).success.value
-
-          val mockSessionRepository = mock[SessionRepository]
-
-          when(mockSessionRepository.set(any())) thenReturn Future.successful(true)
-
-          val application = applicationBuilder(userAnswers = Some(userAnswers))
-            .overrides(
-              bind[Navigation].toInstance(new FakeNavigation(onwardRoute)),
-              bind[SessionRepository].toInstance(mockSessionRepository)
-            )
-            .build()
-
-          running(application) {
-            val request =
-              FakeRequest(POST, hasCorrectGoodsRoute)
-                .withFormUrlEncodedBody(("value", ""))
-
-            val boundForm = form.bind(Map("value" -> ""))
-
-            val view = application.injector.instanceOf[HasCorrectGoodsView]
-
-            val result = route(application, request).value
-
-            status(result) mustEqual BAD_REQUEST
-            contentAsString(result) mustEqual view(
-              boundForm,
-              commodity,
-              onSubmitAction,
-              NormalMode,
-              Some(testRecordId)
-            )(
-              request,
-              messages(application)
-            ).toString
-          }
-        }
-
-      }
-
-    }
-
     "For update journey" - {
       lazy val hasCorrectGoodsUpdateRoute =
-        routes.HasCorrectGoodsController.onPageLoadUpdate(NormalMode, testRecordId).url
+        controllers.hasCorrectGoods.routes.UpdateController.onPageLoad(NormalMode, testRecordId).url
       lazy val onSubmitAction: Call       =
-        routes.HasCorrectGoodsController.onSubmitUpdate(NormalMode, testRecordId)
+        controllers.hasCorrectGoods.routes.UpdateController.onSubmit(NormalMode, testRecordId)
       val page: QuestionPage[Boolean]     = HasCorrectGoodsCommodityCodeUpdatePage(testRecordId)
       val updatePage                      = CommodityCodeUpdatePage(testRecordId)
